+++ backend/app/Http/Controllers/PodcastController.php
@@ -11,7 +11,7 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Requests\UpdatePodcastRequest;
-use App\Http\Requests\StorePodcastRequest;   
+use App\Http\Requests\StorePodcastRequest;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -28,8 +28,8 @@ class PodcastController extends Controller
     public function index()
     {
         $podkasts = Podcast::join('user_metadata', 'podcasts.author_id', '=', 'user_metadata.user_id')
-                ->select('podcasts.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname')
-                ->get();
+            ->select('podcasts.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname')
+            ->get();
         return response()->json($podkasts);
     }
 
@@ -58,6 +58,7 @@ class PodcastController extends Controller
      * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s или Y-m-d).
      * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s или Y-m-d).
      * @urlParam updDate string Дата обновления (формат: Y-m-d).
+     * @urlParam operator string Логический оператор для условий поиска ('and' или 'or').
      * 
      * @param \Illuminate\Http\Request $request
      * @return mixed|\Illuminate\Http\JsonResponse
@@ -86,6 +87,8 @@ class PodcastController extends Controller
         $updDate = $request->query('updDate');
         $crtDate = $request->query('crtDate');
 
+        $operator = $request->query('operator', 'and');
+
         $query = Podcast::query();
 
         if ($withAuthors) {
@@ -108,14 +111,29 @@ class PodcastController extends Controller
             }
         } elseif ($podcastId) {
             $query->where('id', $podcastId);
+            $podcast = $query->first(); // Получаем первый подкаст, соответствующий запросу
+            if ($podcast) {
+                $podcast->increment('views'); // Инкрементируем счетчик просмотров
+            }
         }
 
 
         if (!empty($searchFields) && !empty($searchValues)) {
-            foreach ($searchFields as $index => $field) {
-                $value = $searchValues[$index] ?? null;
-                if ($value) {
-                    $query->where($field, 'LIKE', '%' . $value . '%');
+            if ($operator === 'or') {
+                $query->where(function ($query) use ($searchFields, $searchValues) {
+                    foreach ($searchFields as $index => $field) {
+                        $value = $searchValues[$index] ?? null;
+                        if ($value) {
+                            $query->orWhere($field, 'LIKE', '%' . $value . '%');
+                        }
+                    }
+                });
+            } else {
+                foreach ($searchFields as $index => $field) {
+                    $value = $searchValues[$index] ?? null;
+                    if ($value) {
+                        $query->where($field, 'LIKE', '%' . $value . '%');
+                    }
                 }
             }
         }
@@ -152,7 +170,7 @@ class PodcastController extends Controller
         if ($crtDate) {
             $query->whereDate('created_at', '=', $crtDate);
         }
-    
+
         if ($updDate) {
             $query->whereDate('updated_at', '=', $updDate);
         }
@@ -171,7 +189,7 @@ class PodcastController extends Controller
         return $this->successResponse($podcasts->items(), $paginationData, 200);
     }
 
-/**
+    /**
      * Parses the date from the given input.
      * Supports both Y-m-d H:i:s and Y-m-d formats.
      * 
@@ -208,28 +226,19 @@ class PodcastController extends Controller
      */
     public function store(StorePodcastRequest $request)
     {
-        try {
-            if (!Auth::user()->can('create', Podcast::class)) {
-                throw new AccessDeniedHttpException('You do not have permission to create a podcast');
-            }
-
-            $this->validateRequest($request, $request->rules());
-
-            $podcast = Podcast::create(array_merge($request->validated(), [
-                'status' => 'moderating',
-                'author_id' => Auth::id(),
-            ]));
-            return $this->successResponse(['podcast' => $podcast], 'Podcast created successfully', 200);
-        } catch (AccessDeniedHttpException $e) {
-            return $this->handleException($e);
+        if (!Auth::user()->can('create', Podcast::class)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
         }
+
+        $podcast = Podcast::create($request->validated() + [
+            'status' => 'moderating',
+            'author_id' => Auth::id(),
+        ]);     
+
+        return $this->successResponse(['podcast' => $podcast], 'Podcast created successfully', 200);
     }
 
-
-  
-
-
-   /**
+    /**
      * Обновить
      * 
      * Обновление подкаста
@@ -247,25 +256,23 @@ class PodcastController extends Controller
      */
     public function update(UpdatePodcastRequest $request, $id)
     {
-        try{
+        $podcast = Podcast::find($id);
 
-            $podcast = Podcast::findOrFail($id);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
+        }
 
-            if(!Auth::user()->can('update', $podcast)) {
-                throw new AccessDeniedHttpException('You do not have permission to update this podcast');
-            }
+        if (!Auth::user()->can('update', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
 
-            $this->validateRequest($request, $request->rules());
-            $validatedData = $request->validated();
-            $podcast->update($validatedData);
+        $validatedData = $request->validated();
+        $podcast->update($validatedData);
 
-            return $this->successResponse(['podcast' => $podcast], 'Podcast updated successfully', 200);
-        } catch (AccessDeniedHttpException | ModelNotFoundException $e) {
-            return $this->handleException($e);
-        } 
+        return $this->successResponse(['podcast' => $podcast], 'Podcast updated successfully', 200);
     }
 
-     /**
+    /**
      * Обновить
      * 
      * Обновление статуса подкаста
@@ -276,28 +283,24 @@ class PodcastController extends Controller
      */
     public function updateStatus(int $id, Request $request): \Illuminate\Http\JsonResponse
     {
-        try {
-            $newStatus = $request->input('status');
-            $podcast = Podcast::findOrFail($id);
+        $newStatus = $request->input('status');
+        $podcast = Podcast::find($id);
 
-            if (!$podcast) {
-                return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
-            }
-
-            if (!Auth::user()->can('updateStatus', $podcast)) {
-                throw new AccessDeniedHttpException('You do not have permission to update the status of this podcast');
-            }
-
-            if (!in_array($newStatus, Podcast::STATUSES)) {
-                return $this->errorResponse('Invalid status entered', [], 404);
-            }
-
-            $podcast->update(['status' => $newStatus]);
-
-            return $this->successResponse(['podcasts' => $podcast], 'Podcast status updated successfully', 200);
-        } catch (ModelNotFoundException | AccessDeniedHttpException $e) {
-            return $this->handleException($e);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        if (!Auth::user()->can('updateStatus', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        if (!in_array($newStatus, Podcast::STATUSES)) {
+            return $this->errorResponse('Invalid status entered', [], 404);
+        }
+
+        $podcast->update(['status' => $newStatus]);
+
+        return $this->successResponse(['podcasts' => $podcast], 'Podcast status updated successfully', 200);
     }
 
 
@@ -313,20 +316,20 @@ class PodcastController extends Controller
      */
     public function destroy($id)
     {
-        try{
-            $podcast = Podcast::findOrFail($id);
+        $podcast = Podcast::find($id);
 
-            // Проверка прав пользователя
-            if (!Auth::user()->can('delete', $podcast)) {
-                throw new AccessDeniedHttpException('You do not have permission to delete this podcast');
-            }
-
-            $podcast->delete();
-
-            return $this->successResponse(['podcast' => $podcast], 'Podcast deleted successfully', 200);
-        }catch (Exception | ModelNotFoundException$e) {
-            return $this->handleException($e);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+
+        // Проверка прав пользователя
+        if (!Auth::user()->can('delete', $podcast)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $podcast->delete();
+
+        return $this->successResponse(['podcast' => $podcast], 'Podcast deleted successfully', 200);
     }
 
     /**
@@ -370,24 +373,23 @@ class PodcastController extends Controller
      */
     public function likePodcast(Request $request, int $id): \Illuminate\Http\JsonResponse
     {
-        try {
-            $podcast = Podcast::findOrFail($id);
-            $user = Auth::user();
-
-            $like = $podcast->likes()->where('user_id', $user->id)->first();
-
-            if ($like) {
-                $like->delete();
-                $podcast->decrement('likes');
-                return $this->successResponse(['podcasts' => $podcast], 'Podcast unliked successfully', 200);
-            } else {
-                $podcast->likes()->create(['user_id' => $user->id]);
-                $podcast->increment('likes');
-            }
-
-            return $this->successResponse(['podcasts' => $podcast], 'Podcast liked successfully', 200);
-        } catch (ModelNotFoundException $e) {
-            return $this->handleException($e);
+        $podcast = Podcast::find($id);
+        if (!$podcast) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
+        $user = Auth::user();
+
+        $like = $podcast->likes()->where('user_id', $user->id)->first();
+
+        if ($like) {
+            $like->delete();
+            $podcast->decrement('likes');
+            return $this->successResponse(['podcasts' => $podcast], 'Podcast unliked successfully', 200);
+        } else {
+            $podcast->likes()->create(['user_id' => $user->id]);
+            $podcast->increment('likes');
+        }
+
+        return $this->successResponse(['podcasts' => $podcast], 'Podcast liked successfully', 200);
     }
 }
